pyproject.toml
[project]
name = "jessy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram==3.7.0",
    "alembic==1.14.0",
    "asyncpg==0.30.0",
    "boto3==1.35.3",
    "cachetools==5.4.0",
    "greenlet>=3.2.3",
    "loguru==0.7.2",
    "psycopg2-binary==2.9.10",
    "python-dotenv==1.0.1",
    "sqlalchemy==2.0.36",
]
